requirements.txt
paho_mqtt==2.1.0
pyserial==3.5
pytest==8.3.2
pytest-html==4.1.1
pytest-metadata==3.1.1
RPi.GPIO==0.7.1
pytest-json-report==1.5.0
